+++ hydrate-orchestrator/main.go
@@ -28,6 +28,7 @@ type HydrateOrchestrator struct {
 	DeploymentBranch string
 	Event            EventType
 	DotFirestartr    *dagger.Directory
+	GhCliVersion     string
 }
 
 func New(
@@ -63,6 +64,11 @@ func New(
 	// +required
 	dotFirestartr *dagger.Directory,
 
+	//Gh CLI Version
+	// +optional
+	// +default="v2.66.1"
+	ghCliVersion string,
+
 ) *HydrateOrchestrator {
 	return &HydrateOrchestrator{
 		Repo:             repo,
@@ -74,5 +80,6 @@ func New(
 		AuthDir:          authDir,
 		Event:            event,
 		DotFirestartr:    dotFirestartr,
+		GhCliVersion:     ghCliVersion,
 	}
 }